+++ packages/orm/angel_migration_runner/lib/src/mariadb/table.dart
@@ -6,6 +6,11 @@ import 'package:charcode/ascii.dart';
 abstract class MariaDbGenerator {
   static String columnType(MigrationColumn column) {
     var str = column.type.name;
+
+    // Map timestamp time to datetime
+    if (column.type == ColumnType.timeStamp) {
+      str = ColumnType.dateTime.name;
+    }
     if (column.type.hasLength) {
       return '$str(${column.length})';
     } else {